test: import old module catch errors
